+++ src/taskAutom/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "7.16.7"
+__version__ = "7.16.8"
 __author__ = 'Lucas Aimaretto'